annotation - Refactor datetime to int method to util function.
